create game evaluator
